num shares to int everywhere; working on creating lots on sell processing

+++ src/cg/BuyTrade.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 public class BuyTrade extends Trade {
 
 	public BuyTrade(int tradeId, cg.SimpleDate date, Trade.Type tradeType,
-			String ticker, long numShares, float sharePrice, BigDecimal comm,
+			String ticker, int numShares, float sharePrice, BigDecimal comm,
 			Trade.SpecialInstruction instruction, String note) {
 
 		super(tradeId, date, tradeType, ticker, numShares, sharePrice, comm,
@@ -29,7 +29,7 @@ public class BuyTrade extends Trade {
 		return ((float) numShares) * ((float) sharePrice) + comm.floatValue();
 	}
 
-	public float getBasis(long numShares) {
+	public float getBasis(int numShares) {
 		return ((float) numShares) / ((float) this.numShares) * getBasis();
 	}
